Fix error type for GetKeyFromWallet function

+++ src/wallet/wallet.h
@@ -59,7 +59,7 @@ bool HasWallets();
 std::vector<std::shared_ptr<CWallet>> GetWallets();
 std::shared_ptr<CWallet> GetWallet(const std::string& name);
 std::shared_ptr<CWallet> LoadWallet(interfaces::Chain& chain, const WalletLocation& location, std::string& error, std::string& warning);
-std::array<uint8_t, 32> GetKeyFromWallets(rust::Vec<uint8_t> input);
+std::array<uint32_t, 32> GetKeyFromWallets(rust::Vec<uint8_t> input);
 
 CKey GetWalletsKey(CKeyID const & keyid);
